Move method to show permissions dialog to JsonApi and fixh bug
