+++ android-json-form-wizard/src/main/java/com/vijay/jsonwizard/interfaces/JsonApi.java
@@ -107,4 +107,6 @@ public interface JsonApi {
     String getConfirmCloseTitle();
 
     void setConfirmCloseTitle(String title);
+
+    void showPermissionDeniedDialog();
 }
